+++ XOrm/context_commit_test.go
@@ -20,23 +20,23 @@ import (
 // TestContextCommit 测试提交操作。
 func TestContextCommit(t *testing.T) {
 	t.Run("Setup", func(t *testing.T) {
-		defer initializeCommit(XPrefs.Asset())
+		defer initContextCommit(XPrefs.Asset())
 
-		initializeCommit(XPrefs.New())
-		assert.Equal(t, runtime.NumCPU(), commitQueueCount, fmt.Sprintf("默认的提交队列数量应当为 %v", runtime.NumCPU()))
-		assert.Equal(t, 100000, commitQueueCapacity, "默认的提交批次数量应当为 100000")
+		initContextCommit(XPrefs.New())
+		assert.Equal(t, runtime.NumCPU(), contextCommitQueueCount, fmt.Sprintf("默认的提交队列数量应当为 %v。", runtime.NumCPU()))
+		assert.Equal(t, 100000, contextCommitQueueCapacity, "默认的提交批次数量应当为 100000。")
 
-		initializeCommit(XPrefs.New().Set(preferencesCommitQueueCount, 20).Set(preferencesCommitQueueCapacity, 1000))
-		assert.Equal(t, 20, commitQueueCount, "设置的提交队列数量应当为 20")
-		assert.Equal(t, 1000, commitQueueCapacity, "设置的提交批次数量应当为 1000")
+		initContextCommit(XPrefs.New().Set(preferencesContextCommitQueueCount, 20).Set(preferencesContextCommitQueueCapacity, 1000))
+		assert.Equal(t, 20, contextCommitQueueCount, "设置的提交队列数量应当为 20。")
+		assert.Equal(t, 1000, contextCommitQueueCapacity, "设置的提交批次数量应当为 1000。")
 
-		assert.Equal(t, 20, len(commitQueues), "提交队列数量应当为 20")
-		assert.Equal(t, 1000, cap(commitQueues[0]), "提交队列容量应当为 1000")
-		assert.Equal(t, 20, len(commitSetupSig), "初始化信号数量应当为 20")
-		assert.Equal(t, 20, len(commitFlushWait), "提交信号数量应当为 20")
+		assert.Equal(t, 20, len(contextCommitQueues), "提交队列数量应当为 20。")
+		assert.Equal(t, 1000, cap(contextCommitQueues[0]), "提交队列容量应当为 1000。")
+		assert.Equal(t, 20, len(contextCommitSetupSig), "初始化信号数量应当为 20。")
+		assert.Equal(t, 20, len(contextCommitFlushWait), "提交信号数量应当为 20。")
 
 		Close()
-		assert.Equal(t, int32(1), atomic.LoadInt32(&commitCloseSig), "关闭状态标识应该为 1。")
+		assert.Equal(t, int32(1), atomic.LoadInt32(&contextCommitCloseSig), "关闭状态标识应该为 1。")
 	})
 
 	t.Run("Pool", func(t *testing.T) {
@@ -47,14 +47,14 @@ func TestContextCommit(t *testing.T) {
 			go func(i int) {
 				defer wg.Done()
 
-				bobj := commitBatchPool.Get().(*commitBatch)
+				bobj := contextCommitBatchPool.Get().(*commitBatch)
 				bobj.reset()
 				assert.Nil(t, bobj.tag, "commitBatch 对象的 tag 字段应为 nil。")
 				assert.Equal(t, 0, bobj.stime, "commitBatch 对象的 time 字段应为 0。")
 				assert.Nil(t, bobj.objects, "commitBatch 对象的 objs 字段应为 nil。")
 				assert.Nil(t, bobj.prehandler, "commitBatch 对象的 prehandler 字段应为 nil。")
 				assert.Nil(t, bobj.posthandler, "commitBatch 对象的 posthandler 字段应为 nil。")
-				commitBatchPool.Put(bobj)
+				contextCommitBatchPool.Put(bobj)
 			}(i)
 		}
 		wg.Wait()
@@ -62,11 +62,11 @@ func TestContextCommit(t *testing.T) {
 
 	t.Run("Commit", func(t *testing.T) {
 		defer ResetBaseTest()
-		defer initializeCommit(XPrefs.Asset())
+		defer initContextCommit(XPrefs.Asset())
 
 		ResetBaseTest()
 		SetupBaseTest(false, true)
-		initializeCommit(XPrefs.Asset())
+		initContextCommit(XPrefs.Asset())
 
 		model := NewTestBaseModel()
 
@@ -80,7 +80,7 @@ func TestContextCommit(t *testing.T) {
 				// 测试创建操作
 				t.Run("Create", func(t *testing.T) {
 					for j := range 10 { // 10 个批次
-						batch := commitBatchPool.Get().(*commitBatch)
+						batch := contextCommitBatchPool.Get().(*commitBatch)
 						for k := range 10 { // 10 个对象
 							data := NewTestBaseModel()
 							data.ID = i*100 + j*10 + k + 1
@@ -99,9 +99,9 @@ func TestContextCommit(t *testing.T) {
 					datas := make([]*TestBaseModel, 0)
 					model.List(&datas, Cond("id >= {0} && id <= {1}", i*100+1, i*100+100))
 
-					assert.Equal(t, 100, len(datas), "创建操作后的对象数量应该为 100")
+					assert.Equal(t, 100, len(datas), "创建操作后的对象数量应该为 100。")
 					for _, result := range datas {
-						assert.Equal(t, "test_create", result.StringVal, "创建操作后的对象的 StringVal 字段应该为 test_create")
+						assert.Equal(t, "test_create", result.StringVal, "创建操作后的对象的 StringVal 字段应该为 test_create。")
 					}
 				})
 
@@ -110,7 +110,7 @@ func TestContextCommit(t *testing.T) {
 					datas := make([]*TestBaseModel, 0)
 					model.List(&datas, Cond("id >= {0} && id <= {1}", i*100+1, i*100+100))
 
-					batch := commitBatchPool.Get().(*commitBatch)
+					batch := contextCommitBatchPool.Get().(*commitBatch)
 					for _, data := range datas {
 						data.StringVal = "test_update"
 						sobj := sessionObjectPool.Get().(*sessionObject)
@@ -123,7 +123,7 @@ func TestContextCommit(t *testing.T) {
 					datas = make([]*TestBaseModel, 0)
 					model.List(&datas, Cond("id >= {0} && id <= {1}", i*100+1, i*100+100)) // 重新读取数据
 					for _, result := range datas {
-						assert.Equal(t, "test_update", result.StringVal, "更新操作后的对象的 StringVal 字段应该为 test_update")
+						assert.Equal(t, "test_update", result.StringVal, "更新操作后的对象的 StringVal 字段应该为 test_update。")
 					}
 				})
 
@@ -132,7 +132,7 @@ func TestContextCommit(t *testing.T) {
 					datas := make([]*TestBaseModel, 0)
 					model.List(&datas, Cond("id >= {0} && id <= {1}", i*100+1, i*100+100))
 
-					batch := commitBatchPool.Get().(*commitBatch)
+					batch := contextCommitBatchPool.Get().(*commitBatch)
 					for j := range len(datas) / 2 {
 						data := datas[j]
 						sobj := sessionObjectPool.Get().(*sessionObject)
@@ -145,12 +145,12 @@ func TestContextCommit(t *testing.T) {
 
 					datas = make([]*TestBaseModel, 0)
 					model.List(&datas, Cond("id >= {0} && id <= {1}", i*100+1, i*100+100)) // 重新读取数据
-					assert.Equal(t, 50, len(datas), "删除操作后的对象数量应该为 50")
+					assert.Equal(t, 50, len(datas), "删除操作后的对象数量应该为 50。")
 				})
 
 				// 测试清除操作
 				t.Run("Clear", func(t *testing.T) {
-					batch := commitBatchPool.Get().(*commitBatch)
+					batch := contextCommitBatchPool.Get().(*commitBatch)
 					sobj := sessionObjectPool.Get().(*sessionObject)
 					sobj.ptr = model
 					sobj.clear = Cond("id >= {0} && id <= {1}", i*100+1, i*100+100)
@@ -160,12 +160,12 @@ func TestContextCommit(t *testing.T) {
 
 					datas := make([]*TestBaseModel, 0)
 					model.List(&datas, Cond("id >= {0} && id <= {1}", i*100+1, i*100+100)) // 重新读取数据
-					assert.Equal(t, 0, len(datas), "清除作后的对象数量应该为 0")
+					assert.Equal(t, 0, len(datas), "清除作后的对象数量应该为 0。")
 				})
 
 				// 测试前后置处理器回调
 				t.Run("Handler", func(t *testing.T) {
-					batch := commitBatchPool.Get().(*commitBatch)
+					batch := contextCommitBatchPool.Get().(*commitBatch)
 					sobj := sessionObjectPool.Get().(*sessionObject)
 					sobj.ptr = model
 					batch.objects = append(batch.objects, sobj)
@@ -184,11 +184,11 @@ func TestContextCommit(t *testing.T) {
 					batch.submit()
 					Flush() // 等待提交完成
 
-					assert.Equal(t, batch, prebatch, "前置处理器回调的批次对象应该与提交时的对象相同")
-					assert.Equal(t, sobj, preobj, "前置处理器回调的提交对象应该与提交时的对象相同")
+					assert.Equal(t, batch, prebatch, "前置处理器回调的批次对象应该与提交时的对象相同。")
+					assert.Equal(t, sobj, preobj, "前置处理器回调的提交对象应该与提交时的对象相同。")
 
-					assert.Equal(t, batch, postbatch, "后置处理器回调的批次对象应该与提交时的对象相同")
-					assert.Equal(t, sobj, postobj, "后置处理器回调的提交对象应该与提交时的对象相同")
+					assert.Equal(t, batch, postbatch, "后置处理器回调的批次对象应该与提交时的对象相同。")
+					assert.Equal(t, sobj, postobj, "后置处理器回调的提交对象应该与提交时的对象相同。")
 				})
 			}(i, t)
 		}
@@ -197,16 +197,16 @@ func TestContextCommit(t *testing.T) {
 
 	t.Run("Metrics", func(t *testing.T) {
 		defer ResetBaseTest()
-		defer initializeCommit(XPrefs.Asset())
+		defer initContextCommit(XPrefs.Asset())
 
 		ResetBaseTest()
 		SetupBaseTest(false, true)
-		initializeCommit(XPrefs.Asset())
+		initContextCommit(XPrefs.Asset())
 
 		gid := goid.Get()
-		queueID := max(int(gid)%commitQueueCount, 0)
+		queueID := max(int(gid)%contextCommitQueueCount, 0)
 
-		batch := commitBatchPool.Get().(*commitBatch)
+		batch := contextCommitBatchPool.Get().(*commitBatch)
 		for i := range 100 { // 100 个对象
 			data := NewTestBaseModel()
 			data.ID = i + 1
@@ -219,25 +219,25 @@ func TestContextCommit(t *testing.T) {
 		}
 		batch.submit()
 
-		assert.Equal(t, 100, int(testutil.ToFloat64(commitGauges[queueID])), "指定队列 %v 等待提交的对象数量应当为 100。", queueID)
-		assert.Equal(t, 100, int(testutil.ToFloat64(commitGauge)), "所有队列等待提交的对象数量应当为 100。")
+		assert.Equal(t, 100, int(testutil.ToFloat64(contextCommitGauges[queueID])), "指定队列 %v 等待提交的对象数量应当为 100。", queueID)
+		assert.Equal(t, 100, int(testutil.ToFloat64(contextCommitGauge)), "所有队列等待提交的对象数量应当为 100。")
 
 		Flush() // 等待提交完成
 
-		assert.Equal(t, 0, int(testutil.ToFloat64(commitGauges[queueID])), "指定队列 %v 等待提交的对象数量应当为 0。", queueID)
-		assert.Equal(t, 0, int(testutil.ToFloat64(commitGauge)), "所有队列等待提交的对象数量应当为 0。")
+		assert.Equal(t, 0, int(testutil.ToFloat64(contextCommitGauges[queueID])), "指定队列 %v 等待提交的对象数量应当为 0。", queueID)
+		assert.Equal(t, 0, int(testutil.ToFloat64(contextCommitGauge)), "所有队列等待提交的对象数量应当为 0。")
 
-		assert.Equal(t, 100, int(testutil.ToFloat64(commitCounters[queueID])), "指定队列 %v 已经提交的对象总数应当为 100。", queueID)
-		assert.Equal(t, 100, int(testutil.ToFloat64(commitCounter)), "所有队列已经提交的对象总数应当为 100。")
+		assert.Equal(t, 100, int(testutil.ToFloat64(contextCommitCounters[queueID])), "指定队列 %v 已经提交的对象总数应当为 100。", queueID)
+		assert.Equal(t, 100, int(testutil.ToFloat64(contextCommitCounter)), "所有队列已经提交的对象总数应当为 100。")
 	})
 
 	t.Run("Flush", func(t *testing.T) {
 		defer ResetBaseTest()
-		defer initializeCommit(XPrefs.Asset())
+		defer initContextCommit(XPrefs.Asset())
 
 		ResetBaseTest()
 		SetupBaseTest(false, true)
-		initializeCommit(XPrefs.Asset())
+		initContextCommit(XPrefs.Asset())
 
 		model := NewTestBaseModel()
 
@@ -248,7 +248,7 @@ func TestContextCommit(t *testing.T) {
 			go func(i int) {
 				defer wg.Done()
 
-				batch := commitBatchPool.Get().(*commitBatch)
+				batch := contextCommitBatchPool.Get().(*commitBatch)
 				data := NewTestBaseModel()
 				data.ID = i + 1
 				data.StringVal = "test_flush"