+++ XOrm/context_incre.go
@@ -13,7 +13,7 @@ import (
 	"github.com/petermattis/goid"
 )
 
-// Increment 获取并自增指定列的最大值。model 参数为要操作的数据模型，必须实现 IModel 接口。
+// Incre 获取并自增指定列的最大值。model 参数为要操作的数据模型，必须实现 IModel 接口。
 // columnAndDelta 为可变参数，支持多种组合：无参数时自增主键且增量为 1；一个参数时，若为字符串则
 // 指定列名且增量为 1，若为整数则使用主键并指定增量；两个参数时，第一个为列名（字符串），第二个为增量（整数）。
 //
@@ -23,33 +23,33 @@ import (
 //
 // 需要注意的是，缓存的最大值在程序重启后会重置。
 // 该函数是线程安全的，可以确保单实例内的数据唯一性。
-func Increment(model IModel, columnAndDelta ...any) int {
+func Incre(model IModel, columnAndDelta ...any) int {
 	cacheDumpWait.Wait()
 
 	gid := goid.Get()
 	ctx := getContext(gid)
 	if ctx == nil {
-		XLog.Critical("XOrm.Increment: context was not found: %v", XLog.Caller(1, false))
+		XLog.Critical("XOrm.Incre: context was not found: %v", XLog.Caller(1, false))
 		return -1
 	}
 	meta := getModelMeta(model)
 	if meta == nil {
-		XLog.Critical("XOrm.Increment: model of %v was not registered: %v", model.ModelUnique(), XLog.Caller(1, false))
+		XLog.Critical("XOrm.Incre: model of %v was not registered: %v", model.ModelUnique(), XLog.Caller(1, false))
 		return -1
 	}
 	if !ctx.writable {
-		XLog.Error("XOrm.Increment: context was not writable.")
+		XLog.Error("XOrm.Incre: context was not writable.")
 		return -1
 	}
 	if !meta.writable {
-		XLog.Error("XOrm.Increment: model of %v was not writable.", model.ModelUnique())
+		XLog.Error("XOrm.Incre: model of %v was not writable.", model.ModelUnique())
 		return -1
 	}
 
 	time := XTime.GetMicrosecond()
 	defer func() {
-		ctx.incrementCost += XTime.GetMicrosecond() - time
-		ctx.incrementCount++
+		ctx.increCost += XTime.GetMicrosecond() - time
+		ctx.increCount++
 	}()
 
 	delta := 1
@@ -77,21 +77,21 @@ func Increment(model IModel, columnAndDelta ...any) int {
 		}
 	}
 	if cname == "" {
-		XLog.Error("XOrm.Increment: column was empty: %v", model.ModelUnique())
+		XLog.Error("XOrm.Incre: column was empty: %v", model.ModelUnique())
 		return 0
 	}
 
-	incrementKey := fmt.Sprintf("%v_%v", model.ModelUnique(), cname)
-	if val, ok := globalIncrementMap.Load(incrementKey); !ok {
+	increKey := fmt.Sprintf("%v_%v", model.ModelUnique(), cname)
+	if val, ok := globalIncreMap.Load(increKey); !ok {
 		defer globalIncreMutex.Unlock()
 		globalIncreMutex.Lock()
-		if nval, ok := globalIncrementMap.Load(incrementKey); ok {
+		if nval, ok := globalIncreMap.Load(increKey); ok {
 			return int(atomic.AddInt64(nval.(*int64), int64(delta)))
 		} else {
 			index := model.Max([]string{cname}...)
 			index += delta
 			lindex := int64(index)
-			globalIncrementMap.Store(incrementKey, &lindex)
+			globalIncreMap.Store(increKey, &lindex)
 			return index
 		}
 	} else {