+++ XOrm/context_incre_test.go
@@ -50,22 +50,22 @@ func TestContextIncre(t *testing.T) {
 					contextMap.Store(gid, ctx)
 					defer contextMap.Delete(gid)
 
-					Increment(model)
-					Increment(model, "int_val", 2)
+					Incre(model)
+					Incre(model, "int_val", 2)
 				}(i)
 			}
 			wg.Wait()
 
 			if !test.writable {
-				if _, ok := globalIncrementMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "id")); ok {
-					t.Errorf("只读模式下 id 自增值应当为空")
+				if _, ok := globalIncreMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "id")); ok {
+					t.Errorf("只读模式下 id 自增值应当为空。")
 				}
-				if _, ok := globalIncrementMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "int_val")); ok {
-					t.Errorf("只读模式下 int_val 自增值应当为空")
+				if _, ok := globalIncreMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "int_val")); ok {
+					t.Errorf("只读模式下 int_val 自增值应当为空。")
 				}
 			} else {
-				idIncre, _ := globalIncrementMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "id"))
-				intValIncre, _ := globalIncrementMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "int_val"))
+				idIncre, _ := globalIncreMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "id"))
+				intValIncre, _ := globalIncreMap.Load(fmt.Sprintf("%v_%v", model.ModelUnique(), "int_val"))
 				assert.Equal(t, 1100, int(*(idIncre).(*int64)), "id 的自增值应当为 1100。")
 				assert.Equal(t, 1200, int(*(intValIncre).(*int64)), "int_val 的自增值应当为 1200。")
 			}