+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 更新记录
 
+## [1.0.1] - 
+### 变更
+- 调整函数 XOrm.Register 为短名称 Meta
+- 调整函数 XOrm.Increment 为短名称 Incre
+- 调整配置键 Orm/Commit/Queue 为 XOrm/Context/Commit/Queue
+- 调整配置键 Orm/Commit/Queue/Capacity 为 XOrm/Context/Commit/Capacity
+- 调整指标 xorm_commit_queue 为 xorm_context_commit_queue
+- 调整指标 xorm_commit_total 为 xorm_context_commit_total
+
 ## [1.0.0] - 2025-10-19
 ### 新增
 - 首次发布